requirements.txt
streamlit
build123d
google-genai
python-dotenv
requests
chromadb
sentence-transformers